Split up schema.go
