+++ chunk/schema.go
@@ -1,20 +1,11 @@
 package chunk
 
 import (
-	"bytes"
-	"encoding/base64"
-	"encoding/binary"
-	"encoding/hex"
-	"flag"
 	"fmt"
-	"sort"
-	"strconv"
 	"strings"
 	"time"
 
 	"github.com/prometheus/common/model"
-
-	"github.com/weaveworks/cortex/util"
 )
 
 const (
@@ -61,227 +52,6 @@ type IndexEntry struct {
 	// - If neither is set, must read all keys for that row.
 	RangeValuePrefix []byte
 	RangeValueStart  []byte
-}
-
-// SchemaConfig contains the config for our chunk index schemas
-type SchemaConfig struct {
-	PeriodicTableConfig
-	OriginalTableName string
-
-	// After midnight on this day, we start bucketing indexes by day instead of by
-	// hour.  Only the day matters, not the time within the day.
-	DailyBucketsFrom util.DayValue
-
-	// After this time, we will only query for base64-encoded label values.
-	Base64ValuesFrom util.DayValue
-
-	// After this time, we will read and write v4 schemas.
-	V4SchemaFrom util.DayValue
-
-	// After this time, we will read and write v5 schemas.
-	V5SchemaFrom util.DayValue
-
-	// After this time, we will read and write v6 schemas.
-	V6SchemaFrom util.DayValue
-}
-
-// RegisterFlags adds the flags required to config this to the given FlagSet
-func (cfg *SchemaConfig) RegisterFlags(f *flag.FlagSet) {
-	cfg.PeriodicTableConfig.RegisterFlags(f)
-
-	flag.StringVar(&cfg.OriginalTableName, "dynamodb.original-table-name", "", "The name of the DynamoDB table used before versioned schemas were introduced.")
-	f.Var(&cfg.DailyBucketsFrom, "dynamodb.daily-buckets-from", "The date (in the format YYYY-MM-DD) of the first day for which DynamoDB index buckets should be day-sized vs. hour-sized.")
-	f.Var(&cfg.Base64ValuesFrom, "dynamodb.base64-buckets-from", "The date (in the format YYYY-MM-DD) after which we will stop querying to non-base64 encoded values.")
-	f.Var(&cfg.V4SchemaFrom, "dynamodb.v4-schema-from", "The date (in the format YYYY-MM-DD) after which we enable v4 schema.")
-	f.Var(&cfg.V5SchemaFrom, "dynamodb.v5-schema-from", "The date (in the format YYYY-MM-DD) after which we enable v5 schema.")
-	f.Var(&cfg.V6SchemaFrom, "dynamodb.v6-schema-from", "The date (in the format YYYY-MM-DD) after which we enable v6 schema.")
-}
-
-func (cfg *SchemaConfig) tableForBucket(bucketStart int64) string {
-	if !cfg.UsePeriodicTables || bucketStart < (cfg.PeriodicTableStartAt.Unix()) {
-		return cfg.OriginalTableName
-	}
-	// TODO remove reference to time package here
-	return cfg.TablePrefix + strconv.Itoa(int(bucketStart/int64(cfg.TablePeriod/time.Second)))
-}
-
-type bucketCallback func(from, through uint32, tableName, hashKey string) ([]IndexEntry, error)
-
-func (cfg SchemaConfig) hourlyBuckets(from, through model.Time, userID string, metricName model.LabelValue, callback bucketCallback) ([]IndexEntry, error) {
-	var (
-		fromHour    = from.Unix() / secondsInHour
-		throughHour = through.Unix() / secondsInHour
-		result      = []IndexEntry{}
-	)
-
-	for i := fromHour; i <= throughHour; i++ {
-		relativeFrom := util.Max64(0, int64(from)-(i*millisecondsInHour))
-		relativeThrough := util.Min64(millisecondsInHour, int64(through)-(i*millisecondsInDay))
-		entries, err := callback(uint32(relativeFrom), uint32(relativeThrough), cfg.tableForBucket(i*secondsInHour), fmt.Sprintf("%s:%d:%s", userID, i, metricName))
-		if err != nil {
-			return nil, err
-		}
-		result = append(result, entries...)
-	}
-	return result, nil
-}
-
-func (cfg SchemaConfig) dailyBuckets(from, through model.Time, userID string, metricName model.LabelValue, callback bucketCallback) ([]IndexEntry, error) {
-	var (
-		fromDay    = from.Unix() / secondsInDay
-		throughDay = through.Unix() / secondsInDay
-		result     = []IndexEntry{}
-	)
-
-	for i := fromDay; i <= throughDay; i++ {
-		// The idea here is that the hash key contains the bucket start time (rounded to
-		// the nearest day).  The range key can contain the offset from that, to the
-		// (start/end) of the chunk. For chunks that span multiple buckets, these
-		// offsets will be capped to the bucket boundaries, i.e. start will be
-		// positive in the first bucket, then zero in the next etc.
-		//
-		// The reason for doing all this is to reduce the size of the time stamps we
-		// include in the range keys - we use a uint32 - as we then have to base 32
-		// encode it.
-
-		relativeFrom := util.Max64(0, int64(from)-(i*millisecondsInDay))
-		relativeThrough := util.Min64(millisecondsInDay, int64(through)-(i*millisecondsInDay))
-		entries, err := callback(uint32(relativeFrom), uint32(relativeThrough), cfg.tableForBucket(i*secondsInDay), fmt.Sprintf("%s:d%d:%s", userID, i, metricName))
-		if err != nil {
-			return nil, err
-		}
-		result = append(result, entries...)
-	}
-	return result, nil
-}
-
-// compositeSchema is a Schema which delegates to various schemas depending
-// on when they were activated.
-type compositeSchema struct {
-	schemas []compositeSchemaEntry
-}
-
-type compositeSchemaEntry struct {
-	start model.Time
-	Schema
-}
-
-type byStart []compositeSchemaEntry
-
-func (a byStart) Len() int           { return len(a) }
-func (a byStart) Swap(i, j int)      { a[i], a[j] = a[j], a[i] }
-func (a byStart) Less(i, j int) bool { return a[i].start < a[j].start }
-
-func newCompositeSchema(cfg SchemaConfig) (Schema, error) {
-	schemas := []compositeSchemaEntry{
-		{0, v1Schema(cfg)},
-	}
-
-	if cfg.DailyBucketsFrom.IsSet() {
-		schemas = append(schemas, compositeSchemaEntry{cfg.DailyBucketsFrom.Time, v2Schema(cfg)})
-	}
-
-	if cfg.Base64ValuesFrom.IsSet() {
-		schemas = append(schemas, compositeSchemaEntry{cfg.Base64ValuesFrom.Time, v3Schema(cfg)})
-	}
-
-	if cfg.V4SchemaFrom.IsSet() {
-		schemas = append(schemas, compositeSchemaEntry{cfg.V4SchemaFrom.Time, v4Schema(cfg)})
-	}
-
-	if cfg.V5SchemaFrom.IsSet() {
-		schemas = append(schemas, compositeSchemaEntry{cfg.V5SchemaFrom.Time, v5Schema(cfg)})
-	}
-
-	if cfg.V6SchemaFrom.IsSet() {
-		schemas = append(schemas, compositeSchemaEntry{cfg.V6SchemaFrom.Time, v6Schema(cfg)})
-	}
-
-	if !sort.IsSorted(byStart(schemas)) {
-		return nil, fmt.Errorf("schemas not in time-sorted order")
-	}
-
-	return compositeSchema{schemas}, nil
-}
-
-func (c compositeSchema) forSchemas(from, through model.Time, callback func(from, through model.Time, schema Schema) ([]IndexEntry, error)) ([]IndexEntry, error) {
-	if len(c.schemas) == 0 {
-		return nil, nil
-	}
-
-	// first, find the schema with the highest start _before or at_ from
-	i := sort.Search(len(c.schemas), func(i int) bool {
-		return c.schemas[i].start > from
-	})
-	if i > 0 {
-		i--
-	} else {
-		// This could happen if we get passed a sample from before 1970.
-		i = 0
-		from = c.schemas[0].start
-	}
-
-	// next, find the schema with the lowest start _after_ through
-	j := sort.Search(len(c.schemas), func(j int) bool {
-		return c.schemas[j].start > through
-	})
-
-	min := func(a, b model.Time) model.Time {
-		if a < b {
-			return a
-		}
-		return b
-	}
-
-	start := from
-	result := []IndexEntry{}
-	for ; i < j; i++ {
-		nextSchemaStarts := model.Latest
-		if i+1 < len(c.schemas) {
-			nextSchemaStarts = c.schemas[i+1].start
-		}
-
-		// If the next schema starts at the same time as this one,
-		// skip this one.
-		if nextSchemaStarts == c.schemas[i].start {
-			continue
-		}
-
-		end := min(through, nextSchemaStarts-1)
-		entries, err := callback(start, end, c.schemas[i].Schema)
-		if err != nil {
-			return nil, err
-		}
-
-		result = append(result, entries...)
-		start = nextSchemaStarts
-	}
-
-	return result, nil
-}
-
-func (c compositeSchema) GetWriteEntries(from, through model.Time, userID string, metricName model.LabelValue, labels model.Metric, chunkID string) ([]IndexEntry, error) {
-	return c.forSchemas(from, through, func(from, through model.Time, schema Schema) ([]IndexEntry, error) {
-		return schema.GetWriteEntries(from, through, userID, metricName, labels, chunkID)
-	})
-}
-
-func (c compositeSchema) GetReadEntriesForMetric(from, through model.Time, userID string, metricName model.LabelValue) ([]IndexEntry, error) {
-	return c.forSchemas(from, through, func(from, through model.Time, schema Schema) ([]IndexEntry, error) {
-		return schema.GetReadEntriesForMetric(from, through, userID, metricName)
-	})
-}
-
-func (c compositeSchema) GetReadEntriesForMetricLabel(from, through model.Time, userID string, metricName model.LabelValue, labelName model.LabelName) ([]IndexEntry, error) {
-	return c.forSchemas(from, through, func(from, through model.Time, schema Schema) ([]IndexEntry, error) {
-		return schema.GetReadEntriesForMetricLabel(from, through, userID, metricName, labelName)
-	})
-}
-
-func (c compositeSchema) GetReadEntriesForMetricLabelValue(from, through model.Time, userID string, metricName model.LabelValue, labelName model.LabelName, labelValue model.LabelValue) ([]IndexEntry, error) {
-	return c.forSchemas(from, through, func(from, through model.Time, schema Schema) ([]IndexEntry, error) {
-		return schema.GetReadEntriesForMetricLabelValue(from, through, userID, metricName, labelName, labelValue)
-	})
 }
 
 // v1Schema was:
@@ -526,22 +296,6 @@ func (labelNameInHashKeyEntries) GetReadMetricLabelValueEntries(_, _ uint32, tab
 // v5Entries includes chunk end time in range key - see #298.
 type v5Entries struct{}
 
-func encodeTime(t uint32) []byte {
-	// timestamps are hex encoded such that it doesn't contain null byte,
-	// but is still lexicographically sortable.
-	throughBytes := make([]byte, 4, 4)
-	binary.BigEndian.PutUint32(throughBytes, t)
-	encodedThroughBytes := make([]byte, 8, 8)
-	hex.Encode(encodedThroughBytes, throughBytes)
-	return encodedThroughBytes
-}
-
-func decodeTime(bs []byte) uint32 {
-	buf := make([]byte, 4, 4)
-	hex.Decode(buf, bs)
-	return binary.BigEndian.Uint32(buf)
-}
-
 func (v5Entries) GetWriteEntries(_, through uint32, tableName, hashKey string, labels model.Metric, chunkID string) ([]IndexEntry, error) {
 	chunkIDBytes := []byte(chunkID)
 	encodedThroughBytes := encodeTime(through)
@@ -658,88 +412,4 @@ func (v6Entries) GetReadMetricLabelValueEntries(from, _ uint32, tableName, hashK
 			RangeValueStart: buildRangeKey(encodedFromBytes),
 		},
 	}, nil
-}
-
-func buildRangeKey(ss ...[]byte) []byte {
-	length := 0
-	for _, s := range ss {
-		length += len(s) + 1
-	}
-	output, i := make([]byte, length, length), 0
-	for _, s := range ss {
-		copy(output[i:i+len(s)], s)
-		i += len(s) + 1
-	}
-	return output
-}
-
-func encodeBase64Value(value model.LabelValue) []byte {
-	encodedLen := base64.RawStdEncoding.EncodedLen(len(value))
-	encoded := make([]byte, encodedLen, encodedLen)
-	base64.RawStdEncoding.Encode(encoded, []byte(value))
-	return encoded
-}
-
-func decodeBase64Value(bs []byte) (model.LabelValue, error) {
-	decodedLen := base64.RawStdEncoding.DecodedLen(len(bs))
-	decoded := make([]byte, decodedLen, decodedLen)
-	if _, err := base64.RawStdEncoding.Decode(decoded, bs); err != nil {
-		return "", err
-	}
-	return model.LabelValue(decoded), nil
-}
-
-func parseRangeValue(rangeValue []byte, value []byte) (string, model.LabelValue, bool, error) {
-	components := make([][]byte, 0, 5)
-	i, j := 0, 0
-	for j < len(rangeValue) {
-		if rangeValue[j] != 0 {
-			j++
-			continue
-		}
-
-		components = append(components, rangeValue[i:j])
-		j++
-		i = j
-	}
-
-	switch {
-	case len(components) < 3:
-		return "", "", false, fmt.Errorf("invalid range value: %x", rangeValue)
-
-	// v1 & v2 schema had three components - label name, label value and chunk ID.
-	// No version number.
-	case len(components) == 3:
-		return string(components[2]), model.LabelValue(components[1]), true, nil
-
-	// v3 schema had four components - label name, label value, chunk ID and version.
-	// "version" is 1 and label value is base64 encoded.
-	case bytes.Equal(components[3], rangeKeyV1):
-		labelValue, err := decodeBase64Value(components[1])
-		return string(components[2]), labelValue, false, err
-
-	// v4 schema wrote v3 range keys and a new range key - version 2,
-	// with four components - <empty>, <empty>, chunk ID and version.
-	case bytes.Equal(components[3], rangeKeyV2):
-		return string(components[2]), model.LabelValue(""), false, nil
-
-	// v5 schema version 3 range key is chunk end time, <empty>, chunk ID, version
-	case bytes.Equal(components[3], rangeKeyV3):
-		return string(components[2]), model.LabelValue(""), false, nil
-
-	// v5 schema version 4 range key is chunk end time, label value, chunk ID, version
-	case bytes.Equal(components[3], rangeKeyV4):
-		labelValue, err := decodeBase64Value(components[1])
-		return string(components[2]), labelValue, false, err
-
-	// v6 schema added version 5 range keys, which have the label value written in
-	// to the value, not the range key. So they are [chunk end time, <empty>, chunk ID, version].
-	case bytes.Equal(components[3], rangeKeyV5):
-		labelValue := model.LabelValue(value)
-		return string(components[2]), labelValue, false, nil
-
-	default:
-		return "", model.LabelValue(""), false, fmt.Errorf("unrecognised version: '%v'", string(components[3]))
-	}
-
 }
