+++ chunk/schema_config.go
@@ -1,0 +1,234 @@
+package chunk
+
+import (
+	"flag"
+	"fmt"
+	"sort"
+	"strconv"
+	"time"
+
+	"github.com/prometheus/common/model"
+
+	"github.com/weaveworks/cortex/util"
+)
+
+// SchemaConfig contains the config for our chunk index schemas
+type SchemaConfig struct {
+	PeriodicTableConfig
+	OriginalTableName string
+
+	// After midnight on this day, we start bucketing indexes by day instead of by
+	// hour.  Only the day matters, not the time within the day.
+	DailyBucketsFrom util.DayValue
+
+	// After this time, we will only query for base64-encoded label values.
+	Base64ValuesFrom util.DayValue
+
+	// After this time, we will read and write v4 schemas.
+	V4SchemaFrom util.DayValue
+
+	// After this time, we will read and write v5 schemas.
+	V5SchemaFrom util.DayValue
+
+	// After this time, we will read and write v6 schemas.
+	V6SchemaFrom util.DayValue
+}
+
+// RegisterFlags adds the flags required to config this to the given FlagSet
+func (cfg *SchemaConfig) RegisterFlags(f *flag.FlagSet) {
+	cfg.PeriodicTableConfig.RegisterFlags(f)
+
+	flag.StringVar(&cfg.OriginalTableName, "dynamodb.original-table-name", "", "The name of the DynamoDB table used before versioned schemas were introduced.")
+	f.Var(&cfg.DailyBucketsFrom, "dynamodb.daily-buckets-from", "The date (in the format YYYY-MM-DD) of the first day for which DynamoDB index buckets should be day-sized vs. hour-sized.")
+	f.Var(&cfg.Base64ValuesFrom, "dynamodb.base64-buckets-from", "The date (in the format YYYY-MM-DD) after which we will stop querying to non-base64 encoded values.")
+	f.Var(&cfg.V4SchemaFrom, "dynamodb.v4-schema-from", "The date (in the format YYYY-MM-DD) after which we enable v4 schema.")
+	f.Var(&cfg.V5SchemaFrom, "dynamodb.v5-schema-from", "The date (in the format YYYY-MM-DD) after which we enable v5 schema.")
+	f.Var(&cfg.V6SchemaFrom, "dynamodb.v6-schema-from", "The date (in the format YYYY-MM-DD) after which we enable v6 schema.")
+}
+
+func (cfg *SchemaConfig) tableForBucket(bucketStart int64) string {
+	if !cfg.UsePeriodicTables || bucketStart < (cfg.PeriodicTableStartAt.Unix()) {
+		return cfg.OriginalTableName
+	}
+	// TODO remove reference to time package here
+	return cfg.TablePrefix + strconv.Itoa(int(bucketStart/int64(cfg.TablePeriod/time.Second)))
+}
+
+type bucketCallback func(from, through uint32, tableName, hashKey string) ([]IndexEntry, error)
+
+func (cfg SchemaConfig) hourlyBuckets(from, through model.Time, userID string, metricName model.LabelValue, callback bucketCallback) ([]IndexEntry, error) {
+	var (
+		fromHour    = from.Unix() / secondsInHour
+		throughHour = through.Unix() / secondsInHour
+		result      = []IndexEntry{}
+	)
+
+	for i := fromHour; i <= throughHour; i++ {
+		relativeFrom := util.Max64(0, int64(from)-(i*millisecondsInHour))
+		relativeThrough := util.Min64(millisecondsInHour, int64(through)-(i*millisecondsInDay))
+		entries, err := callback(uint32(relativeFrom), uint32(relativeThrough), cfg.tableForBucket(i*secondsInHour), fmt.Sprintf("%s:%d:%s", userID, i, metricName))
+		if err != nil {
+			return nil, err
+		}
+		result = append(result, entries...)
+	}
+	return result, nil
+}
+
+func (cfg SchemaConfig) dailyBuckets(from, through model.Time, userID string, metricName model.LabelValue, callback bucketCallback) ([]IndexEntry, error) {
+	var (
+		fromDay    = from.Unix() / secondsInDay
+		throughDay = through.Unix() / secondsInDay
+		result     = []IndexEntry{}
+	)
+
+	for i := fromDay; i <= throughDay; i++ {
+		// The idea here is that the hash key contains the bucket start time (rounded to
+		// the nearest day).  The range key can contain the offset from that, to the
+		// (start/end) of the chunk. For chunks that span multiple buckets, these
+		// offsets will be capped to the bucket boundaries, i.e. start will be
+		// positive in the first bucket, then zero in the next etc.
+		//
+		// The reason for doing all this is to reduce the size of the time stamps we
+		// include in the range keys - we use a uint32 - as we then have to base 32
+		// encode it.
+
+		relativeFrom := util.Max64(0, int64(from)-(i*millisecondsInDay))
+		relativeThrough := util.Min64(millisecondsInDay, int64(through)-(i*millisecondsInDay))
+		entries, err := callback(uint32(relativeFrom), uint32(relativeThrough), cfg.tableForBucket(i*secondsInDay), fmt.Sprintf("%s:d%d:%s", userID, i, metricName))
+		if err != nil {
+			return nil, err
+		}
+		result = append(result, entries...)
+	}
+	return result, nil
+}
+
+// compositeSchema is a Schema which delegates to various schemas depending
+// on when they were activated.
+type compositeSchema struct {
+	schemas []compositeSchemaEntry
+}
+
+type compositeSchemaEntry struct {
+	start model.Time
+	Schema
+}
+
+type byStart []compositeSchemaEntry
+
+func (a byStart) Len() int           { return len(a) }
+func (a byStart) Swap(i, j int)      { a[i], a[j] = a[j], a[i] }
+func (a byStart) Less(i, j int) bool { return a[i].start < a[j].start }
+
+func newCompositeSchema(cfg SchemaConfig) (Schema, error) {
+	schemas := []compositeSchemaEntry{
+		{0, v1Schema(cfg)},
+	}
+
+	if cfg.DailyBucketsFrom.IsSet() {
+		schemas = append(schemas, compositeSchemaEntry{cfg.DailyBucketsFrom.Time, v2Schema(cfg)})
+	}
+
+	if cfg.Base64ValuesFrom.IsSet() {
+		schemas = append(schemas, compositeSchemaEntry{cfg.Base64ValuesFrom.Time, v3Schema(cfg)})
+	}
+
+	if cfg.V4SchemaFrom.IsSet() {
+		schemas = append(schemas, compositeSchemaEntry{cfg.V4SchemaFrom.Time, v4Schema(cfg)})
+	}
+
+	if cfg.V5SchemaFrom.IsSet() {
+		schemas = append(schemas, compositeSchemaEntry{cfg.V5SchemaFrom.Time, v5Schema(cfg)})
+	}
+
+	if cfg.V6SchemaFrom.IsSet() {
+		schemas = append(schemas, compositeSchemaEntry{cfg.V6SchemaFrom.Time, v6Schema(cfg)})
+	}
+
+	if !sort.IsSorted(byStart(schemas)) {
+		return nil, fmt.Errorf("schemas not in time-sorted order")
+	}
+
+	return compositeSchema{schemas}, nil
+}
+
+func (c compositeSchema) forSchemas(from, through model.Time, callback func(from, through model.Time, schema Schema) ([]IndexEntry, error)) ([]IndexEntry, error) {
+	if len(c.schemas) == 0 {
+		return nil, nil
+	}
+
+	// first, find the schema with the highest start _before or at_ from
+	i := sort.Search(len(c.schemas), func(i int) bool {
+		return c.schemas[i].start > from
+	})
+	if i > 0 {
+		i--
+	} else {
+		// This could happen if we get passed a sample from before 1970.
+		i = 0
+		from = c.schemas[0].start
+	}
+
+	// next, find the schema with the lowest start _after_ through
+	j := sort.Search(len(c.schemas), func(j int) bool {
+		return c.schemas[j].start > through
+	})
+
+	min := func(a, b model.Time) model.Time {
+		if a < b {
+			return a
+		}
+		return b
+	}
+
+	start := from
+	result := []IndexEntry{}
+	for ; i < j; i++ {
+		nextSchemaStarts := model.Latest
+		if i+1 < len(c.schemas) {
+			nextSchemaStarts = c.schemas[i+1].start
+		}
+
+		// If the next schema starts at the same time as this one,
+		// skip this one.
+		if nextSchemaStarts == c.schemas[i].start {
+			continue
+		}
+
+		end := min(through, nextSchemaStarts-1)
+		entries, err := callback(start, end, c.schemas[i].Schema)
+		if err != nil {
+			return nil, err
+		}
+
+		result = append(result, entries...)
+		start = nextSchemaStarts
+	}
+
+	return result, nil
+}
+
+func (c compositeSchema) GetWriteEntries(from, through model.Time, userID string, metricName model.LabelValue, labels model.Metric, chunkID string) ([]IndexEntry, error) {
+	return c.forSchemas(from, through, func(from, through model.Time, schema Schema) ([]IndexEntry, error) {
+		return schema.GetWriteEntries(from, through, userID, metricName, labels, chunkID)
+	})
+}
+
+func (c compositeSchema) GetReadEntriesForMetric(from, through model.Time, userID string, metricName model.LabelValue) ([]IndexEntry, error) {
+	return c.forSchemas(from, through, func(from, through model.Time, schema Schema) ([]IndexEntry, error) {
+		return schema.GetReadEntriesForMetric(from, through, userID, metricName)
+	})
+}
+
+func (c compositeSchema) GetReadEntriesForMetricLabel(from, through model.Time, userID string, metricName model.LabelValue, labelName model.LabelName) ([]IndexEntry, error) {
+	return c.forSchemas(from, through, func(from, through model.Time, schema Schema) ([]IndexEntry, error) {
+		return schema.GetReadEntriesForMetricLabel(from, through, userID, metricName, labelName)
+	})
+}
+
+func (c compositeSchema) GetReadEntriesForMetricLabelValue(from, through model.Time, userID string, metricName model.LabelValue, labelName model.LabelName, labelValue model.LabelValue) ([]IndexEntry, error) {
+	return c.forSchemas(from, through, func(from, through model.Time, schema Schema) ([]IndexEntry, error) {
+		return schema.GetReadEntriesForMetricLabelValue(from, through, userID, metricName, labelName, labelValue)
+	})
+}
