+++ chunk/schema_test.go
@@ -4,15 +4,12 @@ import (
 	"bytes"
 	"encoding/base64"
 	"fmt"
-	"math"
-	"math/rand"
 	"reflect"
 	"sort"
 	"testing"
 	"time"
 
 	"github.com/prometheus/common/model"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/weaveworks/common/test"
 	"github.com/weaveworks/cortex/util"
@@ -31,149 +28,6 @@ func (mockSchema) GetReadEntriesForMetricLabel(from, through model.Time, userID 
 }
 func (mockSchema) GetReadEntriesForMetricLabelValue(from, through model.Time, userID string, metricName model.LabelValue, labelName model.LabelName, labelValue model.LabelValue) ([]IndexEntry, error) {
 	return nil, nil
-}
-
-func TestSchemaComposite(t *testing.T) {
-	type result struct {
-		from, through model.Time
-		schema        Schema
-	}
-	collect := func(results *[]result) func(from, through model.Time, schema Schema) ([]IndexEntry, error) {
-		return func(from, through model.Time, schema Schema) ([]IndexEntry, error) {
-			*results = append(*results, result{from, through, schema})
-			return nil, nil
-		}
-	}
-	cs := compositeSchema{
-		schemas: []compositeSchemaEntry{
-			{model.TimeFromUnix(0), mockSchema(1)},
-			{model.TimeFromUnix(100), mockSchema(2)},
-			{model.TimeFromUnix(200), mockSchema(3)},
-		},
-	}
-
-	for i, tc := range []struct {
-		cs            compositeSchema
-		from, through int64
-		want          []result
-	}{
-		// Test we have sensible results when there are no schema's defined
-		{compositeSchema{}, 0, 1, []result{}},
-
-		// Test we have sensible results when there is a single schema
-		{
-			compositeSchema{
-				schemas: []compositeSchemaEntry{
-					{model.TimeFromUnix(0), mockSchema(1)},
-				},
-			},
-			0, 10,
-			[]result{
-				{model.TimeFromUnix(0), model.TimeFromUnix(10), mockSchema(1)},
-			},
-		},
-
-		// Test we have sensible results for negative (ie pre 1970) times
-		{
-			compositeSchema{
-				schemas: []compositeSchemaEntry{
-					{model.TimeFromUnix(0), mockSchema(1)},
-				},
-			},
-			-10, -9,
-			[]result{},
-		},
-		{
-			compositeSchema{
-				schemas: []compositeSchemaEntry{
-					{model.TimeFromUnix(0), mockSchema(1)},
-				},
-			},
-			-10, 10,
-			[]result{
-				{model.TimeFromUnix(0), model.TimeFromUnix(10), mockSchema(1)},
-			},
-		},
-
-		// Test we have sensible results when there is two schemas
-		{
-			compositeSchema{
-				schemas: []compositeSchemaEntry{
-					{model.TimeFromUnix(0), mockSchema(1)},
-					{model.TimeFromUnix(100), mockSchema(2)},
-				},
-			},
-			34, 165,
-			[]result{
-				{model.TimeFromUnix(34), model.TimeFromUnix(100) - 1, mockSchema(1)},
-				{model.TimeFromUnix(100), model.TimeFromUnix(165), mockSchema(2)},
-			},
-		},
-
-		// Test we get only one result when two schema start at same time
-		{
-			compositeSchema{
-				schemas: []compositeSchemaEntry{
-					{model.TimeFromUnix(0), mockSchema(1)},
-					{model.TimeFromUnix(10), mockSchema(2)},
-					{model.TimeFromUnix(10), mockSchema(3)},
-				},
-			},
-			0, 165,
-			[]result{
-				{model.TimeFromUnix(0), model.TimeFromUnix(10) - 1, mockSchema(1)},
-				{model.TimeFromUnix(10), model.TimeFromUnix(165), mockSchema(3)},
-			},
-		},
-
-		// Test all the various combination we can get when there are three schemas
-		{
-			cs, 34, 65,
-			[]result{
-				{model.TimeFromUnix(34), model.TimeFromUnix(65), mockSchema(1)},
-			},
-		},
-
-		{
-			cs, 244, 6785,
-			[]result{
-				{model.TimeFromUnix(244), model.TimeFromUnix(6785), mockSchema(3)},
-			},
-		},
-
-		{
-			cs, 34, 165,
-			[]result{
-				{model.TimeFromUnix(34), model.TimeFromUnix(100) - 1, mockSchema(1)},
-				{model.TimeFromUnix(100), model.TimeFromUnix(165), mockSchema(2)},
-			},
-		},
-
-		{
-			cs, 151, 264,
-			[]result{
-				{model.TimeFromUnix(151), model.TimeFromUnix(200) - 1, mockSchema(2)},
-				{model.TimeFromUnix(200), model.TimeFromUnix(264), mockSchema(3)},
-			},
-		},
-
-		{
-			cs, 32, 264,
-			[]result{
-				{model.TimeFromUnix(32), model.TimeFromUnix(100) - 1, mockSchema(1)},
-				{model.TimeFromUnix(100), model.TimeFromUnix(200) - 1, mockSchema(2)},
-				{model.TimeFromUnix(200), model.TimeFromUnix(264), mockSchema(3)},
-			},
-		},
-	} {
-		t.Run(fmt.Sprintf("TestSchemaComposite[%d]", i), func(t *testing.T) {
-			have := []result{}
-			tc.cs.forSchemas(model.TimeFromUnix(tc.from), model.TimeFromUnix(tc.through), collect(&have))
-			if !reflect.DeepEqual(tc.want, have) {
-				t.Fatalf("wrong schemas - %s", test.Diff(tc.want, have))
-			}
-		})
-	}
 }
 
 type ByHashRangeKey []IndexEntry
@@ -525,94 +379,5 @@ func TestSchemaRangeKey(t *testing.T) {
 				require.NoError(t, err)
 			}
 		})
-	}
-}
-
-func TestParseRangeValue(t *testing.T) {
-	// Test we can decode legacy range values
-	for _, c := range []struct {
-		encoded        []byte
-		value, chunkID string
-	}{
-		{[]byte("1\x002\x003\x00"), "2", "3"},
-
-		// version 1 range keys (v3 Schema) base64-encodes the label value
-		{[]byte("toms\x00Y29kZQ\x002:1484661279394:1484664879394\x001\x00"),
-			"code", "2:1484661279394:1484664879394"},
-
-		// version 1 range keys (v4 Schema) doesn't have the label name in the range key
-		{[]byte("\x00Y29kZQ\x002:1484661279394:1484664879394\x001\x00"),
-			"code", "2:1484661279394:1484664879394"},
-
-		// version 2 range keys (also v4 Schema) don't have the label name or value in the range key
-		{[]byte("\x00\x002:1484661279394:1484664879394\x002\x00"),
-			"", "2:1484661279394:1484664879394"},
-
-		// version 3 range keys (v5 Schema) have timestamp in first 'dimension'
-		{[]byte("a1b2c3d4\x00\x002:1484661279394:1484664879394\x003\x00"),
-			"", "2:1484661279394:1484664879394"},
-
-		// version 4 range keys (also v5 Schema) have timestamp in first 'dimension',
-		// base64 value in second
-		{[]byte("a1b2c3d4\x00Y29kZQ\x002:1484661279394:1484664879394\x004\x00"),
-			"code", "2:1484661279394:1484664879394"},
-	} {
-		chunkID, labelValue, _, err := parseRangeValue(c.encoded, nil)
-		require.NoError(t, err)
-		assert.Equal(t, model.LabelValue(c.value), labelValue)
-		assert.Equal(t, c.chunkID, chunkID)
-	}
-}
-
-func TestSchemaTimeEncoding(t *testing.T) {
-	assert.Equal(t, uint32(0), decodeTime(encodeTime(0)), "0")
-	assert.Equal(t, uint32(math.MaxUint32), decodeTime(encodeTime(math.MaxUint32)), "MaxUint32")
-
-	for i := 0; i < 100; i++ {
-		a, b := uint32(rand.Int31()), uint32(rand.Int31())
-
-		assert.Equal(t, a, decodeTime(encodeTime(a)), "a")
-		assert.Equal(t, b, decodeTime(encodeTime(b)), "b")
-
-		if a < b {
-			assert.Equal(t, -1, bytes.Compare(encodeTime(a), encodeTime(b)), "lt")
-		} else if a > b {
-			assert.Equal(t, 1, bytes.Compare(encodeTime(a), encodeTime(b)), "gt")
-		} else {
-			assert.Equal(t, 1, bytes.Compare(encodeTime(a), encodeTime(b)), "eq")
-		}
-	}
-}
-
-func TestSchemaDailyBuckets(t *testing.T) {
-	const (
-		userID     = "0"
-		metricName = model.LabelValue("name")
-		tableName  = "table"
-	)
-	var (
-		config = SchemaConfig{
-			OriginalTableName: tableName,
-		}
-	)
-
-	for _, c := range []struct {
-		from, through model.Time
-	}{
-		{
-			from:    model.TimeFromUnix(0),
-			through: model.TimeFromUnix(2 * 24 * 3600),
-		},
-	} {
-		var i int64
-		_, err := config.dailyBuckets(c.from, c.through, userID, metricName, func(from, through uint32, tableName, hashKey string) ([]IndexEntry, error) {
-			require.True(t, (i*millisecondsInDay)+int64(from) >= int64(c.from), "%d <= %d", (i*millisecondsInDay)+int64(from), int64(c.from))
-			require.True(t, (i*millisecondsInDay)+int64(from) <= int64(c.through), "%d >= %d", (i*millisecondsInDay)+int64(from), int64(c.through))
-			require.True(t, (i*millisecondsInDay)+int64(through) >= int64(c.from), "%d <= %d", (i*millisecondsInDay)+int64(through), int64(c.from))
-			require.True(t, (i*millisecondsInDay)+int64(through) <= int64(c.through), "%d >= %d", (i*millisecondsInDay)+int64(through), int64(c.through))
-			i++
-			return nil, nil
-		})
-		assert.NoError(t, err)
 	}
 }
